% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_diet_matrix.R
\name{compute_diet_matrix}
\alias{compute_diet_matrix}
\title{Computes the posterior diet matrix}
\usage{
compute_diet_matrix(mydata, myres, quant = c(0.025, 0.5, 0.975))
}
\arguments{
\item{mydata}{a list as returned by function \code{\link{prepare_data}}}

\item{myres}{a fit as returned by \code{\link{fit_escroc}}}

\item{quant}{a vector of quantiles}
}
\value{
a  3D array of dim length(quant) x nb species x nb species
}
\description{
Gives the quantiles of posterior diet
}
\examples{
#importing data
data(signature_data)
data(prior_diet_matrix)
data(LOQ)
data(prior_signature_data)
prior_delta <- data.frame(tracer=c("X15N","X13C"),mean=c(3,0),sd=c(1,1))

#check that everything is ok
mydata <- prepare_data(prior_diet_matrix,signature_data,
LOQ,prior_signature_data,prior_delta)

#build the model
mymodel <- building_model(mydata)

#fit the model
myresults <- fit_escroc(mydata, mymodel)

#diet
compute_diet_matrix(mydata, myresults)
}
