% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ena.R
\name{compute_ena}
\alias{compute_ena}
\title{compute ENA indices}
\usage{
compute_ena(mydata, myres, quant = NULL, biomassDet = 0, minTL = 1)
}
\arguments{
\item{mydata}{a list as returned by function \code{\link{prepare_data}}}

\item{myres}{a fit as returned by \code{\link{fit_escroc}}}

\item{quant}{quantiles of interest}

\item{biomassDet}{biomass of detritus}

\item{minTL}{minimum trophic level}
}
\value{
a named array of indices
}
\description{
Computes several ENA indices for an escropath model. ENA indices are based
on formula of Guesnet et al. This function requires enaR package that is
not available on cran anymore but can be found on github
https://github.com/SEELab/enaR
}
\references{

}
