% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_flow_matrix.R
\name{compute_flow_matrix}
\alias{compute_flow_matrix}
\title{Computes the posterior the flow matrix}
\usage{
compute_flow_matrix(mydata, myres, quant = c(0.025, 0.5, 0.975))
}
\arguments{
\item{mydata}{a list as returned by function \code{\link{prepare_data}}}

\item{myres}{a fit as returned by \code{\link{fit_escroc}}}

\item{quant}{a vector of quantiles}
}
\value{
a  3D array of dim length(quant) x nb species x nb species
}
\description{
Gives the quantiles of posterior flow matrix of an ecopath model
}
