% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LOQ}
\alias{LOQ}
\title{Limit of quantifications}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{FOSA}{log10 LOQ FOSA}
  \item{L.PFOS}{log10 LOQ L.PFOS}
  \item{PFNA}{log10 LOQ in PFNA}
  \item{PFOA}{log10 LOQ PFOA}
  \item{PFUnDA}{log10 LOQ PFUnDA}
  \item{X13C}{LOQ is meaningless for Carbon so we put a very low value}
  \item{X15N}{LOQ is meaningless for Nitrogen so we put a very low value}
}

#' @references Munoz et al. (2017) Evidence for the Trophic Transfer of Perfluoroalkylated Substances in a Temperate Macrotidal Estuary. Environmental Science & Technology: 51, 8450-8459}
\source{
\url{https://pubs.acs.org/doi/abs/10.1021/acs.est.7b02399}
}
\usage{
LOQ
}
\description{
Limit of quantifications estimated by Munoz et al. (2017). The number of lines
is similar to the number of lines of \code{\link{signature_data}} and gives for
each individual and each tracer the limit of quantifications of the measures. In
this experiment, the LOQ were similar for all individuals so that lines are all
similar.
}
\keyword{datasets}
