% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myresults}
\alias{myresults}
\title{Example of results}
\format{a mcmc.list}
\usage{
myresults
}
\description{
An example of results send by \code{\link{fit_escroc}} with burnin=1e6, adapt=10000,sample=50000 and thin=150. The thinning
rate was chosen to limit the size of the package.
}
\keyword{datasets}
