% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_delta_tracers.R
\name{plot_delta_tracers}
\alias{plot_delta_tracers}
\title{Plot Magnification / enrichment}
\usage{
plot_delta_tracers(myfit, mydata, trac_names, transfo = list(NULL))
}
\arguments{
\item{myfit}{a fit as returned by \code{\link{fit_escroc}}}

\item{mydata}{a list as returned by \code{\link{prepare_data}}}

\item{trac_names}{a character or vector list containing the tracers
to be plotted}

\item{transfo}{either NULL or a list of functions, on per tracer, to carry transformation if required}
}
\value{
a list of the same length as trac_names with a \code{\link[ggplot2]{ggplot}}
}
\description{
Plot the posterior distribution of magnification or enrichment for specified tracers
}
\examples{
#importing data
data(signature_data)
data(prior_diet_matrix)
data(LOQ)
data(prior_signature_data)
prior_delta <- data.frame(tracer=c("X15N","X13C"),mean=c(3,0),sd=c(1,1))

#check that everything is ok
mydata <- prepare_data(prior_diet_matrix,signature_data,isLeftCensored,
LOQ,prior_signature_data,prior_delta)

#build the model
mymodel <- building_model(mydata)

#fit the model
myresults <- fit_escroc(mydata, mymodel)

#a function to compute TMF ()
transfo <- function (x) 10^x
#plot diet
print(plot_delta_tracers(myresults,mydata,"FOSA",list(transfo)))
}
