% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diet_species.R
\name{plot_diet_species}
\alias{plot_diet_species}
\title{Plot diet}
\usage{
plot_diet_species(myfit, mydata, spec_names)
}
\arguments{
\item{myfit}{a fit as returned by \code{\link{fit_escroc}}}

\item{mydata}{a list as returned by \code{\link{prepare_data}}}

\item{spec_names}{a character or vector list containing the species names
to be plotted}
}
\value{
a list of the same length as spec_names with a \code{\link[ggplot2]{ggplot}}
}
\description{
Plot the diet composition of asked species
}
\examples{
#importing data
data(signature_data)
data(prior_diet_matrix)
data(LOQ)
data(prior_signature_data)
prior_delta <- data.frame(tracer=c("X15N","X13C"),mean=c(3,0),sd=c(1,1))

#check that everything is ok
mydata <- prepare_data(prior_diet_matrix,signature_data,prior_signature_data,isLeftCensored,
LOQ,prior_delta)

#build the model
mymodel <- building_model(mydata)

#fit the model
myresults <- fit_escroc(mydata, mymodel)

#plot diet
print(plot_diet_species(myresults,mydata,"Sole"))
}
