% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_full_diet_matrix.R
\name{plot_full_diet_matrix}
\alias{plot_full_diet_matrix}
\title{plot the full diet matrix}
\usage{
plot_full_diet_matrix(myfit, mydata)
}
\arguments{
\item{myfit}{a fit as returned by \code{\link{fit_escroc}}}

\item{mydata}{a list as returned by \code{\link{prepare_data}}}
}
\value{
nothing
}
\description{
Plot the full diet matrix. Each line corresponds to a predator and its column to the prey.
The color of the circle corresponds to the quantile 50% of the contribution of the prey
to the diet of the predator. The radius of the larger circle (respectively smaller circle)
corresponds to the quantile 97.5% (respectively 2.5%) of the contribution of the prey
to the diet of the predator
}
\examples{
#importing data
data(signature_data)
data(prior_diet_matrix)
data(LOQ)
data(prior_signature_data)
prior_delta <- data.frame(tracer=c("X15N","X13C"),mean=c(3,0),sd=c(1,1))

#check that everything is ok
mydata <- prepare_data(prior_diet_matrix,signature_data,prior_signature_data,isLeftCensored,
LOQ,prior_delta)

#build the model
mymodel <- building_model(mydata)

#fit the model
myresults <- fit_escroc(mydata, mymodel)

#plot the diet matrix
plot_full_diet_matrix(myresults,mydata)
}
