% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prior_signatures_data}
\alias{prior_signatures_data}
\title{Measurement from previous studies}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{Species}{name of the species}
  \item{tracer}{name of the tracer, contaminant of isotopic ratio}
  \item{mean}{mean value observed on the sample (no transformation here)}
  \item{sd}{standard value observed on the sample (no transformation here)}
  \item{n}{number of samples used to estimate mean and sd}
}

#' @references David (2006) Dynamique spatio-temporelle du zooplancton dans l'estuaire de la Gironde et implications au sein du réseau trophique planctonique. Ph.D. Thesis, Univ. de Bordeaux 1, 313pp}
\usage{
prior_signatures_data
}
\description{
This table summarizes measurements made by David (2006). Each line gives
the mean, standard-deviation and number of samples from measurements made
on a give tracer. Values should be log (or log10) transformed if required,
similarly to data in \code{\link{signatures_data}}
}
\keyword{datasets}
