% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signature_data}
\alias{signature_data}
\title{Contaminants and isotopic measures}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{Species}{name of the species}
  \item{FOSA}{log10 concentration in FOSA}
  \item{L.PFOS}{log10 concentration in L.PFOS}
  \item{PFNA}{log10 concentration in PFNA}
  \item{PFOA}{log10 concentration in PFOA}
  \item{PFUnDA}{log10 concentration in PFUnDA}
  \item{X13C}{carbon isotopic ratio}
  \item{X15N}{nitrogen isotopic ratio}
}
}
\source{
\url{https://pubs.acs.org/doi/abs/10.1021/acs.est.7b02399}
}
\usage{
signature_data
}
\description{
Measures of concentrations of different perfluoroalkylated
substances and isotopic ratios on 138 individuals from 16 species
in the Gironde estuary collected (Munoz et al. 2017). NA correspond
to left censored observations
}
\references{
Munoz et al. (2017) Evidence for the Trophic Transfer of Perfluoroalkylated Substances in a Temperate Macrotidal Estuary. Environmental Science & Technology: 51, 8450-8459
}
\keyword{datasets}
