% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_model.R
\name{building_model}
\alias{building_model}
\title{Build the jags model code}
\usage{
building_model(mydata)
}
\arguments{
\item{mydata}{a list data returned by \code{\link{prepare_data}}}
}
\value{
a string corresponding to the jags model
}
\description{
Based on data, this function build the jags model that will be used by runjags.
The model is either an escropath or an escroc model depeding on the data
}
\examples{
#importing data
data(signature_data)
data(prior_diet_matrix)
data(LOQ)
data(prior_signature_data)
prior_delta <- data.frame(tracer=c("X15N","X13C"),mean=c(3,0),sd=c(1,1))

#check that everything is ok
mydata <- prepare_data(prior_diet_matrix,signature_data,
LOQ,prior_signature_data, prior_delta)

#build the model
mymodel <- building_model(mydata)
}
