% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{checking_data}
\alias{checking_data}
\title{Checking data format
check that the data are correctly formated. This function can be called by the user and will be called by \code{\link{prepare_data}}}
\usage{
checking_data(diet_matrix, signatures_data, prior_signatures_data = NULL,
  isLeftCensored, LOQ)
}
\arguments{
\item{diet_matrix}{a square matrix with similar species as colnames and rownames. Species should be consistent with Species in signature_data and prior_signature_data}

\item{signatures_data}{a data frame. First column should be names species and followning columns should correspond to tracers. Data should be transformed first if required. Left censored data and missing data should be denoted NA}

\item{prior_signatures_data}{a data frame that contains data that can be used to build a prior signatures for source species. First column should be names Species and second one tracer. Species and tracers should be consistent with signatures_data. Then, the three following columns correpond to the mean signature, the standard deviation and the number of samples.}

\item{isLeftCensored}{a data frame with the same number of rows as signatures_data and one column per tracer. For each measure from the signatures_data, tells if the observation is left censored (0) or not (1). Put NA for missing data}

\item{LOQ}{same structure as isLeftCensored. Gives the limit of detection of limit of quantification of the corresponding measure}
}
\value{
The function does not return any results but raises an error if a missformat is detected
}
\description{
Checking data format
check that the data are correctly formated. This function can be called by the user and will be called by \code{\link{prepare_data}}
}
\examples{
#importing data
data(signatures_data)
data(prior_diet_matrix)
data(LOQ)
data(prior_signatures_data)

#check that everything is ok
checking_data(prior_diet_matrix,signatures_data,prior_signatures_data,isLeftCensored,LOQ)
}
