% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{checking_data}
\alias{checking_data}
\title{Checking data format}
\usage{
checking_data(prior_diet_matrix, signature_data,
  prior_signature_data = NULL, isLeftCensored, LOQ, prior_delta = NULL)
}
\arguments{
\item{prior_diet_matrix}{a square matrix with similar species as colnames and rownames. Species should be consistent with Species in signature_data and prior_signature_data}

\item{signature_data}{a data frame. First column should be names species and followning columns should correspond to tracers. Data should be transformed first if required. Left censored data and missing data should be denoted NA}

\item{prior_signature_data}{a data frame that contains data that can be used to build a prior signature for source species. First column should be names Species and second one tracer. Species and tracers should be consistent with signature_data. Then, the three following columns correpond to the mean signature, the standard deviation and the number of samples.}

\item{isLeftCensored}{a data frame with the same number of rows as signature_data and one column per tracer. For each measure from the signature_data, tells if the observation is left censored (0) or not (1). Put NA for missing data}

\item{LOQ}{same structure as isLeftCensored. Gives the limit of detection of limit of quantification of the corresponding measure}

\item{prior_delta}{a table that contains prior on magnification/enrichment of tracers per trophic level. First column contains the name of the tracer, second column contains the mean value and third column the standard deviation}
}
\value{
The function does not return any results but raises an error if a missformat is detected
}
\description{
check that the data are correctly formated. This function can be called by the
user and will be called by \code{\link{prepare_data}}
}
\examples{
#importing data
data(signature_data)
data(prior_diet_matrix)
data(LOQ)
data(prior_signature_data)

#check that everything is ok
checking_data(prior_diet_matrix,signature_data,prior_signature_data,
isLeftCensored,LOQ)
}
