% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_init.R
\name{generate_init}
\alias{generate_init}
\title{Generate init
This function is called by \code{\link{fit_escroc}} to
generate relevant initials values for the 3 Markov chains.}
\usage{
generate_init(mydata)
}
\arguments{
\item{mydata}{a list generated by \code{\link{prepare_data}}}
}
\value{
a list of lists of initial values for MCMC chains
}
\description{
Generate init
This function is called by \code{\link{fit_escroc}} to
generate relevant initials values for the 3 Markov chains.
}
